RVM-1063 : Fix wrong return value for opt implementation of d2l for Long.MAX_VALUE as double.
The return value was -1 before the fix and is now the correct 9223372036854775807L.

+++ rvm/src/org/jikesrvm/compilers/opt/lir2mir/ia32/ComplexLIR2MIRExpansion.java
@@ -668,7 +668,7 @@ public abstract class ComplexLIR2MIRExpansion extends IRTools {
       // Value was >= max long
       maxintBB.appendInstruction(CPOS(s, MIR_Move.create(IA32_MOV,
           result.copyRO(),
-          LC((int)Long.MAX_VALUE))));
+          LC(Long.MAX_VALUE))));
       maxintBB.appendInstruction(CPOS(s, MIR_Branch.create(IA32_JMP,
           nextBB.makeJumpTarget())));
       maxintBB.insertOut(nextBB);
